Began styling navbar
