+++ react-app/src/components/NavBar/NavBar.css
@@ -1,0 +1,32 @@
+nav {
+  position: fixed;
+  background-color: white;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  height: 66px;
+  border-bottom: 0.1rem solid #DCDEDD;
+  font-size: 14px;
+  align-items: center;
+  justify-content: space-evenly;
+}
+
+.nav-logo-button{
+  width: 180px;
+  height: 45px;
+  background-image: url("logo.png");
+  background-size: 20%;
+  background-repeat: no-repeat;
+  cursor: pointer;
+  align-items: center;
+  justify-content: space-around;
+  border: none;
+  background-color: transparent;
+}
+
+.nav-logo-text{
+  color: #00CE8E;
+  font-size: 2rem;
+  font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif
+
+}
